tapcli: move client connection functionality to a new file
This change prepares for a future commit that will consolidate
all command-line functionality into a new package.
